requirements.txt
django
djangorestframework
drf-spectacular[sidecar]
